% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_convert.R
\name{as_cal}
\alias{as_cal}
\alias{as_cal.CalDates}
\alias{as_cal.oxcAARCalibratedDatesList}
\alias{as_cal.oxcAARCalibratedDate}
\alias{as_cal.BchronCalibratedDates}
\title{Convert a foreign object to a cal object}
\usage{
as_cal(x)

\method{as_cal}{CalDates}(x)

\method{as_cal}{oxcAARCalibratedDatesList}(x)

\method{as_cal}{oxcAARCalibratedDate}(x)

\method{as_cal}{BchronCalibratedDates}(x)
}
\arguments{
\item{x}{Object from another package to be converted to a \code{cal} object.}
}
\value{
Vector of class \code{c14_cal} (\link{cal}).
}
\description{
\code{as_cal()} converts objects from other packages that represent calibrated
radiocarbon dates to the c14 \code{cal} class.
Methods are currently implemented for:
\itemize{
\item \code{CalDates}: from \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}
\item \code{oxcAARCalibratedDate} and \code{oxcAARCalibratedDatesList}: from \code{\link[oxcAAR:oxcalCalibrate]{oxcAAR::oxcalCalibrate()}}
\item \code{BchronCalibratedDates}: from \code{\link[Bchron:BchronCalibrate]{Bchron::BchronCalibrate()}}
}

These functions are intended for complex S3 objects from other packages.
See \code{\link[=cal]{cal()}} for a more generic constructor, e.g. using a data frame.
}
\seealso{
Other cal class methods: 
\code{\link{cal_metadata}()},
\code{\link{cal}()}

Other c14 conversion functions: 
\code{\link{as.CalDates.cal}()}
}
\concept{c14 conversion functions}
\concept{cal class methods}
