% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_convert.R
\name{as_cal_dist}
\alias{as_cal_dist}
\alias{as_cal_dist.data.frame}
\alias{as_cal_dist.matrix}
\alias{as_cal_dist.CalDates}
\alias{as_cal_dist.oxcAARCalibratedDatesList}
\alias{as_cal_dist.oxcAARCalibratedDate}
\alias{as_cal_dist.BchronCalibratedDates}
\title{Convert a foreign object to a cal_dist object}
\usage{
as_cal_dist(x, ...)

\method{as_cal_dist}{data.frame}(x, ...)

\method{as_cal_dist}{matrix}(x, ...)

\method{as_cal_dist}{CalDates}(x, ...)

\method{as_cal_dist}{oxcAARCalibratedDatesList}(x, ...)

\method{as_cal_dist}{oxcAARCalibratedDate}(x, ..., which = c("prior", "posterior"))

\method{as_cal_dist}{BchronCalibratedDates}(x, ...)
}
\arguments{
\item{x}{Object from another package to be converted to a \code{cal_dist} object.}

\item{...}{Unused}

\item{which}{Which probability distribution to extract from an \code{oxcAARCalibratedDate}:
\code{"prior"} (default) uses raw calibrated probabilities.
\code{"posterior"} uses Bayesian posterior probabilities (error if not available).}
}
\value{
Vector of class \code{c14_cal_dist} (\link{cal_dist}).
}
\description{
\code{as_cal_dist()} converts objects from other packages that represent calibrated
radiocarbon distributions to the c14 \code{cal_dist} class.
Methods are currently implemented for:
\itemize{
\item \code{CalDates}: from \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}
\item \code{oxcAARCalibratedDate} and \code{oxcAARCalibratedDatesList}: from \code{\link[oxcAAR:oxcalCalibrate]{oxcAAR::oxcalCalibrate()}}
\item \code{BchronCalibratedDates}: from \code{\link[Bchron:BchronCalibrate]{Bchron::BchronCalibrate()}}
}

These functions are intended for complex S3 objects from other packages.
The generic constructor \code{\link[=cal_dist]{cal_dist()}} can be used for data frames and other base
structures.
}
\seealso{
Other cal class methods:
\code{\link[=cal_dist]{cal_dist()}}
}
\concept{c14 conversion functions}
\concept{cal class methods}
