% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_calibrate.R
\name{c14_age_norm}
\alias{c14_age_norm}
\title{Generate the normal distribution of a radiocarbon age}
\usage{
c14_age_norm(age, error, resolution = 1, sigma = 5)
}
\arguments{
\item{age}{Radiocarbon age in years.}

\item{error}{Error associated with the radiocarbon age in years.}

\item{resolution}{Desired resolution in years.}

\item{sigma}{Desired sigma range.}
}
\value{
A vector of probability density with years as names.
}
\description{
Generate the normal distribution of a radiocarbon age
}
\examples{
c14_age_norm(10000, 30)
}
