% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_calibrate.R
\name{c14_calibrate}
\alias{c14_calibrate}
\title{Calibrate radiocarbon dates}
\usage{
c14_calibrate(c14_age, c14_error, c14_curve = IntCal20)
}
\arguments{
\item{c14_age}{Vector of uncalibrated radiocarbon ages.}

\item{c14_error}{Vector of standard errors associated with \code{c14_age}.}

\item{c14_curve}{A \link{c14_curve} object or list of \code{c14_curve} objects. See
\link{c14_curves} for a list of the standard curves provided with the packages.
Default: \code{IntCal20}.}
}
\value{
A \code{cal} vector.
}
\description{
Transforms 'raw' radiocarbon ages into a calendar probability distribution
using a calibration curve.
}
\examples{
c14_calibrate(1000, 30)
c14_calibrate(1000, 30, IntCal20)
}
\concept{tidy radiocarbon functions}
