% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_lab_id.R
\name{c14_control_material}
\alias{c14_control_material}
\title{Control radiocarbon sample materials}
\usage{
c14_control_material(
  x,
  thesaurus = c14_material_thesaurus,
  quiet = FALSE,
  warn_unmatched = TRUE
)
}
\arguments{
\item{x}{Vector of radiocarbon sample materials.}

\item{thesaurus}{Thesaurus to use for sample materials. Defaults to the
\link{c14_material_thesaurus} thesaraus included in the package.}

\item{quiet}{Passed to \code{\link[controller:control]{controller::control()}}. Set to \code{TRUE} to suppress
messages about replaced values. Default: \code{FALSE}.}

\item{warn_unmatched}{Passed to \code{\link[controller:control]{controller::control()}}. Set to \code{FALSE} to
suppress warnings about unmatched values. Default: \code{TRUE}.}
}
\value{
Vector the same length as \code{x} with controlled sample materials.
}
\description{
Standardises a vector of radiocarbon sample materials use a thesaurus.
}
