% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.R
\name{c14_curve}
\alias{c14_curve}
\alias{c14_fcurve}
\title{Radiocarbon calibration curve class}
\usage{
c14_curve(
  cal_age = era::yr(),
  c14_age = numeric(),
  c14_error = numeric(),
  d14c = numeric(),
  d14c_error = numeric()
)

c14_fcurve(cal_age = era::yr(), f14c = numeric(), f14c_error = numeric())
}
\arguments{
\item{cal_age}{Vector of calendar ages. If \code{cal_age} is an object of class
\link[era:yr]{era::yr}, its epoch will be respected, otherwise it is assumed to be
\code{"cal BP"}.}

\item{c14_age}{Vector of uncalibrated radiocarbon ages.}

\item{c14_error}{Vector of errors (sigma values) associated with \code{c14_age}.}

\item{d14c}{Vector of Delta-14C values (optional).}

\item{d14c_error}{Vector of errors (sigma values) associated with \code{d14c}
(optional).}

\item{f14c}{Vector of fraction modern (F14C or pMC) values.}

\item{f14c_error}{Vector of errors (sigma values) associated with \code{f14c}.}
}
\value{
Object of class \code{c14_curve}.
}
\description{
The \code{c14_curve} class represents a radiocarbon calibration curve, which
estimates the calendar ages corresponding to a set of radiocarbon
measurements. Usually calibration is performed with either one of the
standard curves included with the package (see \link{c14_curves}) or read in from
a file with \code{\link[=read_14c]{read_14c()}}.

\code{c14_curve()} constructs a a new calibration curve using values in 14C years,
optionally with their associated d14C values.

\code{c14_fcurve()} constructs a curve using fraction modern values (F14C or pMC),
typically used for "post-bomb" calibration.
}
\references{
\insertRef{Stenstrom2011}{c14}
}
\seealso{
Other calibration curve functions: 
\code{\link{read_14c}()}
}
\concept{calibration curve functions}
