% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-curves.R
\docType{data}
\name{c14_curves}
\alias{c14_curves}
\alias{IntCal20}
\alias{SHCal20}
\alias{Marine20}
\alias{IntCal13}
\alias{SHCal13}
\alias{Marine13}
\alias{IntCal09}
\alias{Marine09}
\alias{IntCal04}
\alias{SHCal04}
\alias{Marine04}
\alias{IntCal98}
\alias{Marine98}
\title{Radiocarbon calibration curves}
\format{
A \link{c14_curve} object with the following fields.

For standard curves:
\describe{
\item{cal_age}{Calendar ages with associated era (\link[era:yr]{era::yr} class).}
\item{c14_age}{Uncalibrated radiocarbon ages.}
\item{c14_error}{Errors (sigma values) associated with \code{c14_age}.}
\item{d14c}{Delta-14C values.}
\item{d14c_error}{Errors (sigma values) associated with \code{d14c}.}
}

For post-bomb curves:
\describe{
\item{cal_age}{Calendar ages with their associated era (\link[era:yr]{era::yr} class).}
\item{f14c}{Fraction modern (F14C or pMC) values.}
\item{f14c_error}{Errors (sigma values) associated with \code{f14c}.}
}

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 9501.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 9501.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 5501.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 5141.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 5141.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 4801.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 3522.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 3651.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 3302.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 2482.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 3301.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 1538.

An object of class \code{c14_curve_14c} (inherits from \code{c14_curve}, \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 1569.
}
\source{
\url{http://intcal.org/curves/} and \url{http://calib.org/CALIBomb/}
}
\usage{
IntCal20

SHCal20

Marine20

IntCal13

SHCal13

Marine13

IntCal09

Marine09

IntCal04

SHCal04

Marine04

IntCal98

Marine98
}
\description{
Radiocarbon calibration requires a calibration curve: an estimate of the true
calendar age associated with a radiocarbon measurement. The following curves
are included with the c14 package:

\describe{
\item{IntCal20}{
IntCal Northern Hemisphere curve, 0–55,000 cal BP \insertCite{Reimer2020}{c14}.
}
\item{SHCal20}{
IntCal Southern Hemisphere curve, 0–55,000 cal BP \insertCite{Hogg2020}{c14}.
}
\item{Marine20}{
IntCal marine curve, 0–55,000 cal BP \insertCite{Heaton2020}{c14}.
}
\item{IntCal13, IntCal09, IntCal04, IntCal98}{
Previous IntCal Northern Hemisphere curves
\insertCite{Reimer2013,Reimer2009,Reimer2004,Stuiver1998}{c14}.
Superseded by \code{IntCal20}.
}
\item{SHCal13, SHCal04}{
Previous IntCal Southern Hemisphere curves
\insertCite{Hogg2013,McCormac2004}{c14}.
Superseded by \code{SHCal20}.
}
\item{Marine13, Marine09, Marine04, Marine98}{
Previous IntCal marine curves
\insertCite{Reimer2013,Reimer2009,Hughen2004,Stuiver1998}{c14}.
Superseded by \code{Marine20}.
}
}

Generally, it is recommended to use the latest curve from the IntCal series:
\code{IntCal20} for samples from the northern hemisphere; \code{SHCal20} for samples
from the southern hemisphere; or \code{Marine20} for samples that obtained their
14C in a marine environment.
}
\references{
\insertAllCited{}
}
\seealso{
Other calibration curve functions: 
\code{\link{c14_curve}()},
\code{\link{read_14c}()}
}
\concept{calibration curve functions}
\keyword{datasets}
