% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-constants.R
\docType{data}
\name{c14_decay}
\alias{c14_decay}
\alias{c14_decay_libby}
\alias{c14_decay_cambridge}
\title{Radiocarbon decay constants}
\format{
A numeric of length 1:
\describe{
\item{c14_decay_libby}{The Libby decay constant, \mjeqn{\lambda_L=8033^{-1}}{1/8033}. The standard for calculating the conventional radiocarbon age (CRA).}
\item{c14_decay_cambridge}{The Cambridge decay constant, \mjeqn{\lambda_C=8267^{-1}}{1/8267}.}
}

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\source{
\insertRef{Stenstrom2011}{c14}
}
\usage{
c14_decay_libby

c14_decay_cambridge
}
\description{
\loadmathjax
Decay constants used in the calculation of radiocarbon ages (see \code{\link[=c14_age]{c14_age()}}).
}
\concept{radiocarbon-related constants}
\keyword{datasets}
