% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_lab_id.R
\name{c14_fix_lab_id}
\alias{c14_fix_lab_id}
\alias{c14_is_lab_id}
\title{Detect and fix malformed radiocarbon laboratory identifiers}
\usage{
c14_fix_lab_id(x)

c14_is_lab_id(x)
}
\arguments{
\item{x}{Vector of potentially malformed lab IDs.}
}
\value{
\code{c14_is_lab_id()} returns a logical vector the same length as \code{x}.

\code{c14_fix_lab_id()} returns \code{x} with values replaced to match the conventional
format where possible.
}
\description{
\code{c14_is_lab_id()} tests whether a lab ID matches the conventional format,
e.g. \code{"OxA-1234"} or \code{"OxA 5678"}.

\code{c14_fix_lab_id()} attempts to coerce malformed lab IDs into this format, by
fixing common issues and variant forms.
}
\details{
The regular expression used to test for the conventional format is
\code{"^([[:alpha:]\\(\\)/]{1,8})[ -\\u2010\\u2013_#\\.\\+](.*)$"}. This accepts unusual
but parsable variants such as \code{"Ki(KIEV)-1234"} or \code{"Gif/LSN_5678"}
}
\examples{
# Valid formats
c14_is_lab_id(c("OxA-1234", "OxA 1234", "OxA", "Gif/LSN_5678"))

# Invalid formats
c14_is_lab_id(c("1234", "1234-OxA"))

# Fix
c14_fix_lab_id(c("OxA", "Gif/LSN_5678", "Ki(KIEV)-1234"))
}
