% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c14_lab_code_thesaurus}
\alias{c14_lab_code_thesaurus}
\title{Radiocarbon laboratory code thesaurus}
\format{
A data frame with 688 rows and 2 variables:
\describe{
\item{canon}{Character. The preferred code for the laboratory.}
\item{variant}{Character. Variant codes.}
}
}
\usage{
c14_lab_code_thesaurus
}
\description{
A thesaurus of canonical (matching those given in \link{c14_labs}) radiocarbon
laboratory codes and variants of these codes observed in various published
datasets.
}
\details{
The thesaurus aims to be precise and conservative. That is, canonical forms
are only given for variants when they can be unambiguously interpreted as
that of a known laboratory listed in \link{c14_labs}. It is case sensitive,
because there are laboratory codes that can only be distinguished by
case (see \link{c14_labs} for examples).

Variants that cannot (or have not yet) been unambiguously matched to a
known, standard lab code have a \code{canon} value of \code{NA}.
}
\keyword{datasets}
