% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c14_labs}
\alias{c14_labs}
\title{Radiocarbon laboratories}
\format{
A data frame with 283 rows and 4 variables:
\describe{
\item{lab_code}{Character. Laboratory code, used to identify published
dates as from this lab.}
\item{active}{Logical. \code{FALSE} if the laboratory is "closed, no longer measuring
14C, or operating under a different code".}
\item{lab}{Character. Name of the laboratory.}
\item{country}{Character.}
}
}
\source{
\url{http://radiocarbon.webhost.uits.arizona.edu/node/11}
}
\usage{
c14_labs
}
\description{
A list of known active and defunct radiocarbon laboratories and lab codes,
maintained by the journal \emph{Radiocarbon}. Last updated 9 April 2021.
}
\details{
\code{lab_code} is \emph{almost} unique, but there are exceptions, and sometimes case
is all that distinguishes two labs:
\itemize{
\item \code{"Gd"} (Gliwice) and \code{"GD"} (Gdansk, defunct);
\item \code{"KI"} (Kiel, now \code{"KiA"}) and \code{"Ki"} (Kiev, now \code{"Ki(KIEV)"});
\item \code{"Lu"} (Lund, now \code{"LuS"}, formerly \code{"LuA"}) and \code{"LU"} (St. Petersburg);
\item \code{"P"} (Max Planck) and \code{"P"} (Pennsylvania, defunct);
\item \code{"Pi"} (Pisa, defunct) and \code{"PI"} (Permafrost Institute, defunct);
\item \code{"TKA"} (Univ. of Tokyo Museum) and \code{"TKa"} (Univ. of Tokyo AMS), also \code{"TK"} (Univ. of Tokyo).
}
}
\keyword{datasets}
