% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-thesaurii.R
\docType{data}
\name{c14_material_thesaurus}
\alias{c14_material_thesaurus}
\title{Radiocarbon sample material thesaurus}
\format{
A data frame with 8854 rows and 2 variables:
\describe{
\item{canon}{Character. The preferred description of the sample material.}
\item{variant}{Character. Variant descriptions.}
}
}
\usage{
c14_material_thesaurus
}
\description{
A thesaurus of standardised radiocarbon sample material descriptions and
variants of these materials observed in various published datasets.
}
\author{
Originally compiled by Clemens Schmid, David Matzig, Dirk Seidensticker,
Joe Roe, and Thomas Huet, for the package
\href{https://docs.ropensci.org/c14bazAAR/}{c14bazAAR}.
}
\keyword{datasets}
