% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_lab_id.R
\name{c14_parse_lab_id}
\alias{c14_parse_lab_id}
\title{Parse radiocarbon laboratory identifiers}
\usage{
c14_parse_lab_id(x, fix = TRUE)
}
\arguments{
\item{x}{Vector of lab IDs, e.g. \code{"OxA-1234"}.}

\item{fix}{If \code{TRUE} (the default), attempts to fix malformed lab IDs before
parsing.}
}
\value{
A data frame with two character columns: \code{lab_code} and \code{lab_number}. If \code{x}
is of length one, this is simplified to a named character vector of length 2.
}
\description{
Splits a vector of radiocarbon lab IDs (e.g. \code{"OxA-1234"})
into its constituent parts: the lab code (\code{"OxA"}) and quasi-numeric unique
identifier (\code{"1324"}). If \code{x} contains values that don't fit the standard
format (\code{"OxA-1234"} or \code{"OxA 4567"}), by default the function will attempt
to fix these with \code{\link[=c14_fix_lab_id]{c14_fix_lab_id()}} before parsing.
}
\examples{
c14_parse_lab_id(c("OxA-1234", "Ly 456", "RT 1234-5678"))

# By default, tries to fix malformed values of x before parsing
# Use fix = FALSE for strict parsing
c14_parse_lab_id(c("OxA 1234", "OxA_5678", "Gif/LSN-123"))
c14_parse_lab_id(c("OxA 1234", "OxA_5678", "Gif/LSN-123"), fix = FALSE)

# Single values of x are simplified to a character vector
c14_parse_lab_id("OxA 1234")
}
