% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Calibrated radiocarbon dates}
\usage{
cal(c14_age = double(), c14_error = double(), curve = c14_curve())
}
\arguments{
\item{c14_age}{Vector of uncalibrated radiocarbon ages.}

\item{c14_error}{Vector of standard errors associated with \code{c14_age}.}

\item{curve}{A \code{c14_curve} object or list of \code{c14_curve} objects.
Default: \code{IntCal20}.}
}
\value{
A vector of class \code{cal} (\code{c14_cal}).
}
\description{
The \code{cal} class represents a vector of calibrated radiocarbon dates as a
lazy record. Each element stores the parameters needed to derive a calendar
probability distribution (\code{c14_age}, \code{c14_error}, and \code{c14_curve}), but does
not compute the distribution until requested.
}
\examples{
cal(1000, 30, curve = IntCal20)
cal(c(1000, 2000), c(30, 40), curve = IntCal20)
}
