% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_summary.R
\name{cal_age_range}
\alias{cal_age_range}
\alias{cal_age_min}
\alias{cal_age_max}
\title{Range of calibrated radiocarbon dates}
\usage{
cal_age_range(x, min_pdens = NULL)

cal_age_min(x, min_pdens = NULL)

cal_age_max(x, min_pdens = NULL)
}
\arguments{
\item{x}{A \link{cal} vector of calendar probability distributions.}

\item{min_pdens}{Controls the minimum probability density threshold used to
determine boundaries. The default is \code{pdens > 0} within a plausible range
around the calibrated date. Set \code{min_pdens} to a value between 0 and 1 to
futher constrain the range or set \code{min_pdens = 0} to find the absolute
boundaries on the full range of the calibration curve.}
}
\value{
A data frame with two columns giving the minimum (\code{min}) and maximum
(\code{max}) ages.
}
\description{
Functions for calculating the minimum and maximum ages of a \link{cal} vector.
The range is determined by the probability distribution, with the extent
controlled by the \code{min_pdens} parameter.
}
\examples{
x <- c14_calibrate(c(10000, 9000, 8000), rep(10, 3))

cal_age_min(x)
cal_age_max(x)
cal_age_range(x)
}
\seealso{
Other functions for summarising calibrated radiocarbon dates:
\code{\link[=cal_hdr]{cal_hdr()}},
\code{\link[=cal_point]{cal_point()}},
\code{\link[=cal_probability]{cal_probability()}}
}
\concept{functions for summarising calibrated radiocarbon dates}
