% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_aggregate.R
\name{cal_density}
\alias{cal_density}
\alias{density.c14_cal}
\title{Kernel density estimation of calendar probability distributions}
\usage{
cal_density(x, bw = 30, ..., times = 25, bootstrap = TRUE, strata = NULL)

\method{density}{c14_cal}(x, ...)
}
\arguments{
\item{x}{A \link{cal} vector of calendar probability distributions}

\item{bw}{Kernel bandwidth size passed to \code{\link[stats:density]{stats::density()}}. Can be either
an integer value or a character selection rule, but the latter will likely
result in a different bandwidth being applied to each bootstrapped sample
and therefore should be avoided. Default: \code{30} (years).}

\item{...}{Further arguments passed to \code{\link[stats:density]{stats::density()}}}

\item{times}{Number of bootstrap samples to generate. The default of \code{25} is
suitable for testing but should be set much higher in practice.}

\item{bootstrap}{If \code{TRUE} (the default), randomly resamples with replacement
from \code{x} before each KDE calculation.}

\item{strata}{If not \code{NULL}, KDEs will be weighted to ensure equal
representation of classes specified by a factor or character vector. If
\code{bootstrap = TRUE}, bootstrap samples are generated using stratified
resampling of the same classes.}
}
\value{
The result as a data frame or \link[tibble:tibble]{tibble::tibble} with three columns: \code{age}
(interpolated to common range of \code{x}), \code{.estimate} (mean of the bootstrapped
KDEs), and \code{.error} (standard error of the bootstrapped KDEs).
}
\description{
Aggregates calibrated radiocarbon dates and other calendar probability
distributions using the composite kernel density estimate (cKDE) method
method \insertCite{Brown2017}{c14}. This involves estimating the density of
a set of dates by repeatedly computing the kernel density estimate of a
random sample of ages drawn from their probability distributions.
}
\details{
\code{cal_density(..., bootstrap = TRUE)} performs composite kernel density
estimation with bootstrapping \insertCite{McLaughlin2019}{c14}, where
sampling error. \code{x} is estimated by randomly resampling \code{x} before each KDE
calculation.

See \insertCite{McLaughlin2019;textual}{c14} and
\insertCite{Crema2022;textual}{c14} for discussions of bandwith selection.
}
\examples{
data(shub1_c14)
shub1_cal <- c14_calibrate(shub1_c14$c14_age, shub1_c14$c14_error)
cal_density(shub1_cal)
}
\references{
\insertAllCited{}
}
