% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_dist.R
\name{cal_dist}
\alias{cal_dist}
\alias{cal_dist.c14_cal}
\alias{cal_dist.data.frame}
\alias{cal_dist.default}
\title{Calendar probability distribution}
\usage{
cal_dist(x, ...)

\method{cal_dist}{c14_cal}(x, at = NULL, ...)

\method{cal_dist}{data.frame}(x, ..., .era = era::era("cal BP"))

\method{cal_dist}{default}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \code{cal_dist}. Currently supports \code{cal} and
\code{data.frame} objects.}

\item{...}{Additional arguments passed to methods.}

\item{at}{List of numeric vectors giving the ages over which the distribution
should be calculated. Recycled to \code{length(cal)}. If \code{NULL} (the default)
then a plausible range is estimated for each calibrated date at the native
resolution of the calibration curve (see details).}

\item{.era}{\code{\link[era:era]{era::era()}} object describing the time scale used for ages.
Defaults to calendar years Before Present (\code{era("cal BP")}). Only used by
the \code{data.frame} method when ages are not already \code{era::yr()} vectors.}
}
\value{
A list of data frames with class \code{cal_dist} (\code{c14_cal_dist}). Each element
has two columns: \code{age} and \code{pdens}.
}
\description{
The \code{cal_dist} class represents a vector of calendar probability
distributions.

\code{cal_dist()} is an S3 generic that constructs a new \code{cal_dist} vector:
\itemize{
\item From a \code{cal} object: evaluates the calibrated probability distribution
\item From a \code{data.frame}: constructs from raw age and probability density data
}
}
\details{
The 'plausible range' of a calibrated radiocarbon distribution is estimated
as the portions of the calibration curve that are no more than six times the
combined error of the determination and the curve away from the conventional
radiocarbon age.
}
\examples{
# From data frames:
cal_dist(data.frame(age = era::yr(1:10, "cal BP"), pdens = rep(0.1, 10)))

# From cal objects:
cal <- c14::cal(5000, 10, c14::IntCal20)
cal_dist(cal)
}
\seealso{
Other cal class methods:
\code{\link[=as_cal]{as_cal()}}
}
\concept{cal class methods}
