% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_summary.R
\name{cal_mode}
\alias{cal_mode}
\alias{cal_point}
\alias{cal_median}
\alias{cal_mean}
\alias{cal_local_mode}
\alias{cal_central}
\title{Point estimates of calibrated radiocarbon dates}
\usage{
cal_mode(x, quiet = FALSE)

cal_median(x)

cal_mean(x)

cal_local_mode(x, interval = 0.954)

cal_central(x, interval = 0.954)
}
\arguments{
\item{x}{\code{cal} object. A vector of calibrated radiocarbon dates.}

\item{quiet}{Set \code{quiet = TRUE} to suppress warnings (only used by
\code{\link[=cal_mode]{cal_mode()}}).}

\item{interval}{Numeric. The confidence interval for \code{\link[=cal_local_mode]{cal_local_mode()}} and
\code{\link[=cal_central]{cal_central()}}. Default: 0.954.}
}
\value{
Numeric vector the same length as \code{x} giving the point estimate for each
calibrated date.
}
\description{
Functions for deriving a single point estimate of calendar age from calibrated
radiocarbon dates. Note that none of these methods produce a \emph{good} estimate
\insertCite{Michczynski2007}{c14}, and should only be used as a last resort
if the probability distribution function or an interval estimate is not
appropriate, in which case \code{\link[=cal_mode]{cal_mode()}} (the default) is recommended.
}
\details{
\describe{
\item{\code{cal_mode()}}{Age corresponding to the maximum peak of the probability
distribution. If \code{x} has more than one modal value, the first is returned.}
\item{\code{cal_median()}}{Age corresponding to the median quantile of the
probability distribution.}
\item{\code{cal_mean()}}{Mean age weighted by probability density.}
\item{\code{cal_local_mode()}}{Age corresponding to the maximum peak of the
probability distribution within the confidence range specified by
\code{interval} (\strong{not yet implemented}).}
\item{\code{cal_central()}}{Age corresponding to the central point of the
probability density within the confidence range specified by \code{interval}
(\strong{not yet implemented}).}
}
}
\examples{
x <- c14_calibrate(10000, 30)

cal_mode(x)
cal_median(x)
cal_mean(x)
}
\references{
\insertAllCited{}
}
\seealso{
Other functions for summarising calibrated radiocarbon dates:
\code{\link[=cal_age_range]{cal_age_range()}},
\code{\link[=cal_hdr]{cal_hdr()}},
\code{\link[=cal_probability]{cal_probability()}}
}
\concept{functions for summarising calibrated radiocarbon dates}
