% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_summary.R
\name{cal_point}
\alias{cal_point}
\title{Point estimates of calibrated radiocarbon dates}
\usage{
cal_point(
  x,
  method = c("mode", "median", "mean", "local_mode", "central"),
  interval = 0.954
)
}
\arguments{
\item{x}{\code{cal} object. A vector of calibrated radiocarbon dates.}

\item{method}{Character. Method of estimation:
\describe{
\item{\code{"mode"} (default)}{age corresponding to the maximum peak of the probability distribution}
\item{\code{"median"}}{age corresponding to the median quantile of the probability distribution}
\item{\code{"mean"}}{mean age weighted by probability density}
\item{\code{"local_mode"}}{age corresponding to the maximum peak of the probability
distribution within the confidence range specified by \code{interval} (\strong{not yet
implemented})}
\item{\code{"central"}}{age corresponding to the central point of the probability
density within the confidence range specified by \code{interval} (\strong{not yet
implemented})}
}}

\item{interval}{Numeric. Only used for \code{method = "local_mode"} and
\code{method = "central"}.}
}
\value{
Numeric vector the same length as \code{x} giving the point estimate for each
calibrated date.
}
\description{
This function implements a number of methods for deriving a single point
estimate of calendar age from calibrated radiocarbon dates. Note that none of
these methods produce a \emph{good} estimate \insertCite{Michczynski2007}{c14},
and should only be used as a last resort if the probability distribution
function or an interval estimate is not appropriate, in which case
\code{method = "mode"} (the default) is recommended.
}
\details{
If \code{x} has more than one modal value, \code{cal_point(x, method = "mode")} will
return the first.
}
\examples{
cal_point(c14_calibrate(10000, 30))
}
\references{
\insertAllCited{}
}
\seealso{
Other functions for summarising calibrated radiocarbon dates: 
\code{\link{cal_age_range}()}
}
\concept{functions for summarising calibrated radiocarbon dates}
