% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-calp.R
\name{calp}
\alias{calp}
\alias{as_calp}
\title{Calendar probability distributions}
\usage{
calp(age = numeric(), pdens = numeric())

as_calp(x, ...)
}
\arguments{
\item{age}{Vector of calendar ages. Either an \link[era:yr]{era::yr} vector with era
specified, or a numeric vector, in which case the era is assumed to be "cal
BP".}

\item{pdens}{Probability densities for each value of \code{age}.}

\item{x}{For \code{as_calp()}, foreign object to convert to a \code{calp} object.}

\item{...}{Arguments passed to other functions (currently unused).}
}
\value{
A data frame with class \code{calp} (\code{c14_calp}) and two columns: \code{age} and \code{pdens}.
}
\description{
The \code{calp} class represents a calibrated radiocarbon or other calendar-based
probability distribution as a data frame with two columns: age and
probability density. It is usually contained within a \link{cal} object and
constructed by a calibration or conversion function.

Use \code{calp()} to construct a new \code{calp} object from a pair of vectors. Use
as \code{as_calp()} to convert an object from another package to the \code{calp} class.
}
\examples{
# Uniform distribution between 1 and 10 BP:
calp(era::yr(1:10, "BP"), rep(0.1, 10))
}
