% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14-package.R
\docType{package}
\name{c14-package}
\alias{c14}
\alias{c14-package}
\title{c14: Tidy Radiocarbon Data}
\description{
S3 classes and functions for tidy radiocarbon data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://c14.joeroe.io}
  \item \url{https://github.com/joeroe/c14}
  \item Report bugs at \url{https://github.com/joeroe/c14/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
