% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14-package.R
\docType{package}
\name{c14-package}
\alias{c14}
\alias{c14-package}
\title{c14: Tidy Radiocarbon Data}
\description{
S3 classes and functions for tidy radiocarbon data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://c14.joeroe.io}
  \item \url{https://github.com/joeroe/c14}
  \item Report bugs at \url{https://github.com/joeroe/c14/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

Other contributors:
\itemize{
  \item Thomas Huet \email{thomashuet7@gmail.com} (\href{https://orcid.org/0000-0002-1112-6122}{ORCID}) [contributor]
  \item David Matzig \email{matzigdavid@gmail.com} (\href{https://orcid.org/0000-0001-7349-5401}{ORCID}) [contributor]
  \item Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [contributor]
  \item Dirk Seidensticker \email{dirk.seidensticker@gmail.com} (\href{https://orcid.org/0000-0002-8155-7702}{ORCID}) [contributor]
}

}
\keyword{internal}
