% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_calibrate.R
\name{c14_age_norm}
\alias{c14_age_norm}
\title{Generate the normal distribution of a radiocarbon age}
\usage{
c14_age_norm(age, error, resolution = 1, sigma = 5)
}
\arguments{
\item{age}{Radiocarbon age(s) in years.}

\item{error}{Error(s) associated with the radiocarbon age in years.}

\item{resolution}{Desired resolution in years.}

\item{sigma}{Desired sigma range.}
}
\value{
A vector of probability density with years as names.
If \code{age} or \code{error} has length >= 1, a list of these vectors.
}
\description{
Generate the normal distribution of a radiocarbon age
}
\details{
\code{age} and \code{error} are recycled to a common length.
}
\examples{
c14_age_norm(10000, 30)
}
