% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_id.R
\name{c14_control_lab_id}
\alias{c14_control_lab_id}
\title{Control radiocarbon laboratory identifiers}
\usage{
c14_control_lab_id(
  x,
  thesaurus = cleanc14::c14_lab_code_thesaurus,
  sep = "-",
  quiet = FALSE,
  warn_unmatched = TRUE
)
}
\arguments{
\item{x}{Vector of radiocarbon laboratory identifiers.}

\item{thesaurus}{Thesaurus to use for lab codes. Defaults to the \link{c14_lab_code_thesaurus}
thesaraus included in the package.}

\item{sep}{Character to use to seperate lab codes and numbers in the result.
Default: \code{"-"}.}

\item{quiet}{Passed to \code{\link[controller:control]{controller::control()}}. Set to \code{TRUE} to suppress
messages about replaced values. Default: \code{FALSE}.}

\item{warn_unmatched}{Passed to \code{\link[controller:control]{controller::control()}}. Set to \code{FALSE} to
suppress warnings about unmatched values. Default: \code{TRUE}.}
}
\value{
Vector the same length as \code{x} with controlled laboratory identifiers.
}
\description{
Standardises a vector of radiocarbon laboratory identifiers to a common
format. Specifically, this \code{c14_control_lab_id()}:
\enumerate{
\item Fixes malformed identifiers with \code{\link[=c14_fix_lab_id]{c14_fix_lab_id()}}
\item Parses the lab code and lab number components (\code{\link[=c14_parse_lab_id]{c14_parse_lab_id()}})
\item Standardises lab codes against a thesaurus, by default \link{c14_lab_code_thesaurus}
\item Reunites lab codes and numbers with a uniform seperator (default: \code{"-"})
}
}
