% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c14_labs}
\alias{c14_labs}
\title{Radiocarbon laboratories}
\format{
A data frame with 297 rows and 4 variables:
\describe{
\item{lab_code}{Character. Laboratory code, used to identify published
dates as from this lab.}
\item{active}{Logical. \code{FALSE} if the laboratory is "closed, no longer measuring
14C, or operating under a different code".}
\item{lab}{Character. Name of the laboratory.}
\item{country}{Character.}
}
}
\usage{
c14_labs
}
\description{
Table of active and defunct radiocarbon laboratories with their standard lab
codes.

Primarily based on the list maintained by the journal \emph{Radiocarbon}
(updated 20 July 2021), with a few additions.
}
\details{
\code{lab_code} is \emph{almost} unique, but there are exceptions, and sometimes case
is all that distinguishes two labs:
\itemize{
\item \code{"Gd"} (Gliwice) and \code{"GD"} (Gdansk, defunct);
\item \code{"KI"} (Kiel, now \code{"KiA"}) and \code{"Ki"} (Kiev, now \code{"Ki(KIEV)"});
\item \code{"Lu"} (Lund, now \code{"LuS"}, formerly \code{"LuA"}) and \code{"LU"} (St. Petersburg);
\item \code{"P"} (Max Planck) and \code{"P"} (Pennsylvania, defunct);
\item \code{"Pi"} (Pisa, defunct) and \code{"PI"} (Permafrost Institute, defunct);
\item \code{"TKA"} (Univ. of Tokyo Museum) and \code{"TKa"} (Univ. of Tokyo AMS), also \code{"TK"} (Univ. of Tokyo).
}
}
\references{
List of known radiocarbon laboratories. \emph{Radiocarbon}. Last updated 20 July
2021. \url{http://radiocarbon.webhost.uits.arizona.edu/node/11}

Wang, C., Lu, H., Zhang, J., Gu, Z. and He, K., 2014. Prehistoric demographic
fluctuations in China inferred from radiocarbon data and their linkage with
climate change over the past 50,000 years. \emph{Quaternary Science Reviews},
98, pp.45-59. doi:10.1016/j.quascirev.2014.05.015
}
\keyword{datasets}
