% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Calibrated radiocarbon dates}
\usage{
cal(..., .era = era::era("cal BP"))
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic-dots}> A set of data frames. Each
should have two columns, the first a vector of calendar ages, and the second
a vector of associated probability densities. If the first column is not an
\code{\link[era:yr]{era::yr()}} vector, it is coerced to one using the time scale specified by
\code{.era}.}

\item{.era}{\code{\link[era:era]{era::era()}} object describing the time scale used for ages.
Defaults to calendar years Before Present (\code{era("cal BP")}). Not used if
the ages specified in \code{...} are already \code{era::yr()} vectors.}
}
\value{
A list of data frames with class \code{cal} (\code{c14_cal}). Each element has two
columns: \code{age} and \code{pdens}.
}
\description{
The \code{cal} class represents a vector of calendar probability distribution;
typically calibrated radiocarbon dates.

\code{cal()} constructs a new \code{cal} vector from a set of data frames containing
the raw probability distributions.
}
\examples{
# Uniform distribution between 1 and 10 BP:
cal(data.frame(age = era::yr(1:10, "cal BP"), pdens = rep(0.1, 10)))
}
