% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_aggregate.R
\name{cal_sum}
\alias{cal_sum}
\title{Sum calendar probability distributions}
\usage{
cal_sum(x, range = cal_age_common(x), ...)
}
\arguments{
\item{x}{A \link{cal} vector of calendar probability distributions}

\item{range}{Range of ages over which to sum. Defaults to the maximum range
of \code{x} at one year resolution.}

\item{...}{Not used}
}
\value{
The summed probability distribution as a \link{cal} vector.
}
\description{
Aggregates a vector of calendar probability distributions (such as calibrated
radiocarbon dates) by summing them over a given range of ages.
}
\details{
This function can be evaluated in parallel using \code{\link[future:plan]{future::plan()}}, which
substantially speeds up the computation for large sets of dates.
}
\examples{
ppnd_cal <- c14_calibrate(ppnd$cra, ppnd$error)
cal_sum(ppnd_cal)
}
\concept{functions for aggregating calendar probability distributions}
