% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanc14-package.R
\docType{package}
\name{cleanc14-package}
\alias{cleanc14}
\alias{cleanc14-package}
\title{cleanc14: Clean Radiocarbon Data}
\description{
Utility functions for cleaning radiocarbon datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/cleanc14}
  \item Report bugs at \url{https://github.com/joeroe/cleanc14/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
